File repository ve routing eklendi

+++ model/file.go
@@ -4,6 +4,6 @@ type File struct {
 	ID        string `json:"id";gorm:"primary_key"`
 	Name      string `json:"name"`
 	Format    string `json:"name"`
-	Size      int32  `json:"size"`
+	Size      int64  `json:"size"`
 	Extension string `json:"extension"`
 }
